project-phase-2-for-realz-this-time-yjs-main/junha/project-phase-i-just-bing-chillin-josh/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
Werkzeug==3.0.2